Moved common funcs to utils
Moved common functions/methods to utils, and have also removed cost
function for time being, as this will be done later.
